[HotsBpHelper] 解析器不再解析比赛事件, 修复config, 提高保存storage频率

+++ HotsBpHelper/Configuration/BpHelperConfigParser.cs
@@ -132,7 +132,7 @@ namespace HotsBpHelper.Configuration
                 sb.AppendLine(WriteConfigurationValue(AutoUploadNewReplayToHotslogsKey, customConfigurationSettings.AutoUploadNewReplayToHotslogs));
                 sb.AppendLine(WriteConfigurationValue(AutoUploadNewReplayToHotsweekKey, customConfigurationSettings.AutoUploadNewReplayToHotsweek));
                 sb.AppendLine(WriteConfigurationValue(DefalutReplayFolderPathKey, customConfigurationSettings.DefalutReplayFolderPath));
-                sb.AppendLine(WriteConfigurationValue(UploadStrategyKey, customConfigurationSettings.UploadStrategy));
+                sb.AppendLine(WriteConfigurationValue(UploadStrategyKey, (int)customConfigurationSettings.UploadStrategy));
                 sb.AppendLine(WriteConfigurationValue(MMRAutoCloseTimeKey, customConfigurationSettings.MMRAutoCloseTime));
             }
